refactor: input-t
